requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-engineio==4.2.1
python-socketio==5.4.0
gunicorn==20.1.0
gevent==21.12.0
gevent-websocket==0.10.1
Werkzeug==2.0.1
click==7.1.2
greenlet==1.1.2
